requirements.txt
flask
flask-cors
anthropic
requests
python-dotenv